+++ inc/Autoloader.php
@@ -11,6 +11,10 @@ declare( strict_types = 1 );
 
 namespace AdvancedScreenOptions;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 /**
  * Class - Autoloader
  */
@@ -68,7 +72,7 @@ final class Autoloader {
 			add_action(
 				$hook,
 				static function (): void {
-					$error_message = __( 'Advanced Screen Options: The Composer autoloader was not found. If you installed the plugin from the GitHub source code, make sure to run `composer install`.', 'screen-options' );
+					$error_message = __( 'Advanced Screen Options: The Composer autoloader was not found. If you installed the plugin from the GitHub source code, make sure to run `composer install`.', 'advanced-screen-options' );
 
 					if ( defined( 'WP_DEBUG' ) && WP_DEBUG ) {
 						error_log( esc_html( $error_message ) ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log -- This is a development notice.