+++ inc/Main.php
@@ -9,6 +9,10 @@ declare( strict_types = 1 );
 
 namespace AdvancedScreenOptions;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 use AdvancedScreenOptions\Contracts\Traits\Singleton;
 
 /**