+++ inc/Modules/Post_Types/Abstract_Post_Type.php
@@ -7,6 +7,10 @@
 
 namespace AdvancedScreenOptions\Modules\Post_Types;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 use AdvancedScreenOptions\Contracts\Interfaces\Registrable;
 
 /**