+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 1.0.0
 
 - Initial public release
+
+## 1.1.0
+- Bug fixes and performance improvements
+- Update: Bump WordPress version compatibility to 7.0
+- Updated text domain from screen-options to advanced-screen-options
+- Refactored code
+- Update pot file